requirements.txt
streamlit
openai
pymupdf
python-dotenv
pydantic